HuBMAP Datasets

datasets returns the details available datasets, ordered by last modified dates

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

dataset_detail() takes a unique dataset_id and returns details about one specified dataset as a tibble

Usage

datasets(size = 10000L, from = 0L)

datasets_default_columns(as = c("tibble", "character"))

dataset_detail(uuid)

Arguments

size

integer(1) number of maximum results to return; The default (10000) is meant to be large enough to return all results.

from

integer(1) number of number of results to skip, defaulting to 0.

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Dataset UUID string. This is expected to be a 32-digit hex number.

Value

*_columns() returns a named list name containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

datasets(size = 100, from = 200)
#> # A tibble: 100 × 9
#>    uuid             hubmap_id data_types dataset_type donor.hubmap_id group_name
#>    <chr>            <chr>     <chr>      <chr>        <chr>           <chr>     
#>  1 dbc5beb00b135b9… HBM892.D… image_pyr… Auto-fluore… HBM889.KSKV.978 Vanderbil…
#>  2 a9503f75064c50b… HBM999.D… image_pyr… Auto-fluore… HBM285.XQXL.628 Vanderbil…
#>  3 1e6fd664bc0f5f9… HBM673.F… image_pyr… Auto-fluore… HBM938.LVRS.434 Vanderbil…
#>  4 82cbdc9a6eef422… HBM332.F… image_pyr… Auto-fluore… HBM687.KPKM.763 Vanderbil…
#>  5 ebec43b391090f6… HBM364.M… image_pyr… Auto-fluore… HBM322.KQBK.747 Vanderbil…
#>  6 6344050b82c0b5d… HBM575.L… image_pyr… Auto-fluore… HBM395.HCTC.292 Vanderbil…
#>  7 b444fe229aa2173… HBM238.J… image_pyr… Auto-fluore… HBM398.QCMJ.558 Vanderbil…
#>  8 647d7cf8d574dd9… HBM396.Z… image_pyr… Auto-fluore… HBM253.KBSM.226 Vanderbil…
#>  9 102da8a31557148… HBM277.N… AF         Auto-fluore… HBM357.GMBX.758 Vanderbil…
#> 10 5863df933c2d6e9… HBM466.H… image_pyr… Auto-fluore… HBM259.SNDB.787 Vanderbil…
#> # ℹ 90 more rows
#> # ℹ 3 more variables: last_modified_timestamp <date>, status <chr>, organ <chr>
datasets_default_columns()
#> # A tibble: 9 × 1
#>   columns                
#>   <chr>                  
#> 1 uuid                   
#> 2 hubmap_id              
#> 3 group_name             
#> 4 dataset_type           
#> 5 data_types             
#> 6 organ                  
#> 7 status                 
#> 8 last_modified_timestamp
#> 9 donor.hubmap_id        

uuid <- "7754aa5ebde628b5e92705e33e74a4ef"
dataset_detail(uuid)
#> # A tibble: 1 × 31
#>   ancestor_ids ancestors  contains_human_genetic_sequen…¹ created_by_user_disp…²
#>   <list>       <list>     <lgl>                           <chr>                 
#> 1 <chr [4]>    <list [4]> FALSE                           HuBMAP Process        
#> # ℹ abbreviated names: ¹​contains_human_genetic_sequences,
#> #   ²​created_by_user_displayname
#> # ℹ 27 more variables: created_by_user_email <chr>, created_timestamp <dbl>,
#> #   creation_action <chr>, data_access_level <chr>, data_types <list>,
#> #   dataset_info <chr>, dataset_type <chr>, descendant_ids <list>,
#> #   descendants <list>, display_subtype <chr>, donor <list>, entity_type <chr>,
#> #   files <list>, group_name <chr>, group_uuid <chr>, hubmap_id <chr>, …