HuBMAP Donors

donors() returns details about available samples, ordered by last modified dates

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

donor_detail() takes a unique donor_id and returns details about one specified sample as a tibble

donor_derived() takes a unique donor_id and returns the derived dataset or/and sample details.

donor_metadata() takes a unique donor_id and returns the metadata of the donor.

Usage

donors()

donors_default_columns(as = c("tibble", "character"))

donor_detail(uuid)

donor_derived(uuid, entity_type = c("Dataset", "Sample"))

donor_metadata(uuid)

Arguments

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Donor UUID string. This is expected to be a 32-digit hex number.

entity_type

character(1) selected derived entity type. One of "Sample" or "Dataset" (default).

Value

*_columns() returns a named list name containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

donors()
#> # A tibble: 304 × 8
#>    uuid                 hubmap_id group_name Sex     Age `Body Mass Index` Race 
#>    <chr>                <chr>     <chr>      <chr> <dbl>             <dbl> <chr>
#>  1 7dee9126a98c1e0d244… HBM994.T… Universit… Male   20                23.8 White
#>  2 9a365990ee2f87e6ae6… HBM945.N… Universit… Fema…  42                37.9 Blac…
#>  3 dc32c49414c71ac8f75… HBM697.M… Vanderbil… Male   70                24.8 White
#>  4 dab3ca946f147660e64… HBM453.W… Stanford … Fema…  24.5              20.5 Asian
#>  5 12327363809f68135c9… HBM285.X… Vanderbil… Male   66                23.7 White
#>  6 a9c2ea2bd3106f142ed… HBM994.N… Universit… Fema…  25                30.2 White
#>  7 6a152306bc81179d7a1… HBM954.J… Stanford … Fema…  21.2              34   Blac…
#>  8 e727257a20a08346765… HBM852.D… Universit… Male   25                20.9 Othe…
#>  9 25fd10f300c96818d34… HBM845.M… General E… Male   38                21.8 Unkn…
#> 10 1e0b527a43693605045… HBM735.Q… Universit… Male   67                32.1 White
#> # ℹ 294 more rows
#> # ℹ 1 more variable: last_modified_timestamp <chr>
donors_default_columns()
#> # A tibble: 8 × 1
#>   columns                
#>   <chr>                  
#> 1 hubmap_id              
#> 2 uuid                   
#> 3 group_name             
#> 4 Sex                    
#> 5 Age                    
#> 6 Body Mass Index        
#> 7 Race                   
#> 8 last_modified_timestamp

uuid <- "1dcde05aea3509b2cf89a41ceb3d700f"
donor_detail(uuid)
#> # A tibble: 1 × 21
#>   ancestor_ids ancestors  created_by_user_displayname created_by_user_email     
#>   <list>       <list>     <chr>                       <chr>                     
#> 1 <list [0]>   <list [0]> Jamie Allen                 jamie.l.allen@vanderbilt.…
#> # ℹ 17 more variables: created_timestamp <dbl>, data_access_level <chr>,
#> #   descendant_ids <list>, descendants <list>, description <chr>,
#> #   display_subtype <chr>, entity_type <chr>, group_name <chr>,
#> #   group_uuid <chr>, hubmap_id <chr>, immediate_ancestor_ids <list>,
#> #   immediate_descendant_ids <list>, index_version <chr>,
#> #   last_modified_timestamp <dbl>, metadata <list>, protocol_url <chr>,
#> #   uuid <chr>
uuid <- "1dcde05aea3509b2cf89a41ceb3d700f"
donor_derived(uuid, "Sample")
#> # A tibble: 7 × 3
#>   uuid                             organ          derived_dataset_count
#>   <chr>                            <chr>                          <int>
#> 1 146c221bca3759ffa64cd01bc7069cf3 Kidney (Right)                    16
#> 2 09af77a53cd7d037b40e36abf378b0d7 Kidney (Right)                     4
#> 3 bf1ef1fc97437e6808e4658f7200e247 Kidney (Right)                     4
#> 4 27485b21a2285dbcae38f2d56d429dec Kidney (Right)                    12
#> 5 1441346ca08aadbb631539663ab1874a Kidney (Right)                     2
#> 6 e97fd3e2b6dae2238cfe052a91409f40 Kidney (Right)                     2
#> 7 92cd878bf5be78131c6a3f701342a3ac Kidney (Right)                     8

uuid <- "d37df2cad4e80dc368763caefccf7140"
donor_metadata(uuid)
#> # A tibble: 8 × 2
#>   Key             Value           
#>   <chr>           <chr>           
#> 1 Age             "49 years"      
#> 2 Body Mass Index "31.2 kg/m2"    
#> 3 Cause of Death  "Gunshot wound "
#> 4 Death Event     "Suicide "      
#> 5 Height          "182.88 cm"     
#> 6 Race            "White "        
#> 7 Sex             "Male "         
#> 8 Weight          "104.55 kg"