HuBMAP Provenance

HuBMAP Provenance

Usage

uuid_provenance(uuid)

Arguments

uuid

character(1) corresponding to the HuBMAP Dataset or Sample UUID string. This is expected to be a 32-digit hex number.

Value

uuid_provenance() takes a Dataset or Sample UUID, and returns the provenance details as a list of characters, from the most recent ancestor to the furthest ancestor. The entity type of ancestors include Dataset, Sample, and Donor. Note that there is no ancestor for Donor entity such that the returned result will be an empty list.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

uuid <- "acf816bed56dd6ba55da82960d937ec4"
uuid_provenance(uuid)
#> [[1]]
#> [1] "69e5acda3ceefb02ef4363be64958c53, Dataset"
#> 
#> [[2]]
#> [1] "8332815c55f72e2aec5610dcb4eab9d7, Sample"
#> 
#> [[3]]
#> [1] "507bb20f36adc8a513fe31a913bcfde6, Sample"
#> 
#> [[4]]
#> [1] "eed96170f42554db84d97d1652bb23ef, Sample"
#> 
#> [[5]]
#> [1] "1628b6f7eb615862322d6274a6bc9fa0, Donor"
#>