Publication Information Summaries from Publication UUID

publication_information() queries the HuBMAP publication for title, abstract, manuscript,corresponding authors, dataset types, and organ.

publication_information() organizes the returned tibble from publication_information() in a more legible format.

Usage

publication_information(uuid)

# S3 method for class 'publication_information'
print(x, ...)

Arguments

uuid

character(1) corresponding to the HuBMAP Publication UUID string. This is expected to be a 32-digit hex number.

x

an object of class publication_information, the result of a call to publication_information().

...

additional arguments, required to conform with the print generic but not used.

Value

publication_information() returns a tibble, and columns containing information about the publication. The tibble is of class publication_information and is printed in character.

print.publication_information() is invoked automatically when the result of publication_information() is displayed for it's side effect of displaying the object.

Examples

uuid <- "e8338966c69e759157d6c6ad24847989"
publication_information(uuid)
#> Title
#>  Organ Mapping Antibody Panels (OMAPs): A community resource for standardized multiplexed tissue imaging
#> Abstract
#>  Multiplexed antibody-based imaging enables the detailed characterization of molecular and cellular organization in tissues. Significant advances in the field now allow high-parameter data collection (60+targets); however, considerable expertise and capital are needed to construct the antibody panels employed by these methods. Organ Mapping Antibody Panels (OMAPs) are community-validated resources that save time and money, increase reproducibility, accelerate discovery and support the construction of a Human Reference Atlas.
#> Manuscript
#>  - Nature Methods: https://doi.org/10.1038/s41592-023-01846-7
#> Corresponding Authors
#>  - Andrea J. Radtke 0000-0003-4379-8967
#> Data Types
#>  - Cell DIVE
#> Organs
#>  - Skin