HuBMAP Samples

publications() returns details about available samples, ordered by last modified dates. publication_status == TRUE: peer reviewed publication; publication_status == FALSE: pre-print publication.

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

publication_data() takes a unique publication_id and returns details about one specified publication.

publication_authors() takes a unique publication_id and returns authors information of one specified collection as a tibble

Usage

publications(size = 10000L, from = 0L)

publications_default_columns(as = c("tibble", "character"))

publication_data(uuid, entity_type = c("Dataset", "Sample", "Donor"))

publication_authors(uuid)

Arguments

size

integer(1) number of maximum results to return; The default (10000) is meant to be large enough to return all results.

from

integer(1) number of number of results to skip, defaulting to 0.

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Publication UUID string. This is expected to be a 32-digit hex number.

entity_type

character(1) selected derived entity type. One of "Sample", "Donor" or "Dataset".

Value

*_columns() returns a named list name

containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

publications(size = 3, from = 2)
#> # A tibble: 3 × 7
#>   uuid      hubmap_id last_modified_timest…¹ publication_date publication_status
#>   <chr>     <chr>     <date>                 <chr>            <lgl>             
#> 1 2ced91fd… HBM885.X… 2023-07-18             2021-10-18       FALSE             
#> 2 e8338966… HBM695.W… 2023-07-18             2023-07-19       TRUE              
#> 3 dd0e7aea… HBM599.D… 2023-07-18             2023-07-19       TRUE              
#> # ℹ abbreviated name: ¹​last_modified_timestamp
#> # ℹ 2 more variables: publication_venue <chr>, title <chr>
publications_default_columns()
#> # A tibble: 7 × 1
#>   columns                
#>   <chr>                  
#> 1 uuid                   
#> 2 hubmap_id              
#> 3 title                  
#> 4 publication_venue      
#> 5 last_modified_timestamp
#> 6 publication_date       
#> 7 publication_status     

uuid <- "3c7273660cdf9ab91a7901533b2cd9a5"
publication_data(uuid)
#> # A tibble: 1 × 7
#>   uuid     hubmap_id data_types dataset_type status last_modified_timest…¹ organ
#>   <chr>    <chr>     <chr>      <chr>        <chr>  <date>                 <chr>
#> 1 00d9d4b… HBM342.M… SNARE-RNA… RNAseq       Publi… 2022-05-11             Kidn…
#> # ℹ abbreviated name: ¹​last_modified_timestamp

uuid <- "3c7273660cdf9ab91a7901533b2cd9a5"
publication_authors(uuid)
#> # A tibble: 50 × 8
#>    affiliation      first_name is_contact last_name middle_name_or_initial name 
#>    <chr>            <chr>      <chr>      <chr>     <chr>                  <chr>
#>  1 Department of B… Blue       FALSE      Lake      "B."                   Blue…
#>  2 Department of C… Rajasree   FALSE      Menon     ""                     Raja…
#>  3 Department of P… Seth       FALSE      Winfree   ""                     Seth…
#>  4 Department of B… Qiwen      FALSE      Hu        ""                     Qiwe…
#>  5 Department of M… Ricardo    FALSE      Melo Fer… ""                     Rica…
#>  6 Department of B… Kian       FALSE      Kalhor    ""                     Kian…
#>  7 Department of M… Daria      FALSE      Barwinska ""                     Dari…
#>  8 Department of I… Edgar      FALSE      Otto      "A."                   Edga…
#>  9 Department of M… Michael    FALSE      Ferkowicz ""                     Mich…
#> 10 Department of B… Dinh       FALSE      Diep      ""                     Dinh…
#> # ℹ 40 more rows
#> # ℹ 2 more variables: orcid_id <chr>, version <chr>