HuBMAP Samples

samples() returns details about available samples, ordered by last modified dates

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

sample_detail() takes a unique sample_id and returns details about one specified sample as a tibble

sample_derived() takes a unique sample_id and returns the derived dataset or/and sample details.

Usage

samples()

samples_default_columns(as = c("tibble", "character"))

sample_detail(uuid)

sample_derived(uuid, entity_type = c("Dataset", "Sample"))

Arguments

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Sample UUID string. This is expected to be a 32-digit hex number.

entity_type

character(1) selected derived entity type. One of "Sample" or "Dataset" (default).

Value

*_columns() returns a named list name

containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

samples()
#> # A tibble: 1,927 × 7
#>    uuid              hubmap_id donor.hubmap_id group_name last_modified_timest…¹
#>    <chr>             <chr>     <chr>           <chr>      <chr>                 
#>  1 509d2f22c9e9c33e… HBM944.X… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  2 1367446a254479d8… HBM668.F… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  3 6f700af74a252c99… HBM363.L… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  4 6874561561f77469… HBM435.T… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  5 f0905232556b1608… HBM396.V… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  6 1cfe67afe97094ac… HBM355.M… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  7 23820292ab47f319… HBM236.X… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  8 038e4fcefc00f7ec… HBM595.S… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  9 e3b74d607c3b20c9… HBM876.S… HBM477.DBJN.952 Universit… 2024-05-29            
#> 10 e4ffe5a1de1e63b3… HBM867.F… HBM568.HLJW.252 Universit… 2024-05-29            
#> # ℹ 1,917 more rows
#> # ℹ abbreviated name: ¹​last_modified_timestamp
#> # ℹ 2 more variables: sample_category <chr>, organ <chr>
samples_default_columns()
#> # A tibble: 7 × 1
#>   columns                
#>   <chr>                  
#> 1 uuid                   
#> 2 hubmap_id              
#> 3 group_name             
#> 4 sample_category        
#> 5 organ                  
#> 6 last_modified_timestamp
#> 7 donor.hubmap_id        

uuid <- "d3525d35f6d5ee3dc3186613b0ab1762"
sample_detail(uuid)
#> # A tibble: 1 × 25
#>   ancestor_ids ancestors  created_by_user_displayname created_by_user_email
#>   <list>       <list>     <chr>                       <chr>                
#> 1 <chr [3]>    <list [3]> Marda Jorgensen             marda@ufl.edu        
#> # ℹ 21 more variables: created_timestamp <dbl>, data_access_level <chr>,
#> #   descendant_ids <list>, descendants <list>, display_subtype <chr>,
#> #   donor <list>, entity_type <chr>, group_name <chr>, group_uuid <chr>,
#> #   hubmap_id <chr>, immediate_ancestors <list>, immediate_descendants <list>,
#> #   index_version <chr>, lab_tissue_sample_id <chr>,
#> #   last_modified_timestamp <dbl>, origin_samples <list>, protocol_url <chr>,
#> #   rui_location <chr>, sample_category <chr>, submission_id <chr>, …
uuid <- "3e7dc14313262af577f686dcb09f5119"
sample_derived(uuid, "Sample")
#> # A tibble: 1 × 5
#>   uuid                    hubmap_id sample_category last_modified_timest…¹ organ
#>   <chr>                   <chr>     <chr>           <chr>                  <chr>
#> 1 63ef1040ae30473778dc4b… HBM976.W… section         2024-04-09             Brain
#> # ℹ abbreviated name: ¹​last_modified_timestamp