HuBMAP Collections

collections() returns details about available collections, ordered by last modified dates

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

collection_detail() takes a unique collection_id and returns details about one specified collection as a tibble

collection_contacts() takes a unique collection_id and returns contacts information of one specified collection as a tibble

collection_data() takes a unique collection_id and returns related datasets of one specified collection as a tibble

collection_contributors() takes a unique collection_id and returns contributors information of one specified collection as a tibble

Usage

collections()

collections_default_columns(as = c("tibble", "character"))

collection_detail(uuid)

collection_contacts(uuid)

collection_data(uuid)

collection_contributors(uuid)

Arguments

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Collection UUID string. This is expected to be a 32-digit hex number.

Value

*_columns() returns a named list name containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

collections()
#> # A tibble: 18 × 4
#>    uuid                             hubmap_id       last_modified_timest…¹ title
#>    <chr>                            <chr>           <chr>                  <chr>
#>  1 3ae4ddfc175d768af5526a010bfe95aa HBM585.QPDV.454 2023-03-27             "Spa…
#>  2 074df152c7d3dc2e09ec2400e947bc0c HBM692.JRZB.356 2022-10-27             "Hig…
#>  3 c87fa4ec3aa87af19a9ad29d2e2b4c5c HBM776.RGSW.867 2022-09-02             "\"A…
#>  4 34b068d4a926f77fd98b3d968b6c172f HBM426.WQXQ.244 2022-04-20             "Hum…
#>  5 4964d24bbc6668a72c4cbb5e0393a6bc HBM925.SGXL.596 2021-05-19             "HuB…
#>  6 6a6efd0c1a2681dc7d2faab8e4ab0bca HBM876.XNRH.336 2020-09-28             "Mul…
#>  7 4b87c7596e1f4b4443612862fc562ff1 HBM748.JFJF.982 2020-09-28             "Mul…
#>  8 eb063a0ed8a68117b17345415ab4dd3c HBM554.MXBG.848 2020-09-28             "Mul…
#>  9 a320abcf56241be2e5e453401efeaaf3 HBM556.TLMK.545 2020-09-28             "Mul…
#> 10 3695ab25f42ee161fd05e75c4688847e HBM496.TWCK.764 2020-09-28             "Mul…
#> 11 90399057f6ff0ff18d4c46a378f0b069 HBM789.JXCT.358 2020-09-28             "Mul…
#> 12 590b0485a196956284b8f3344276bc50 HBM524.VZTB.679 2020-09-28             "Mul…
#> 13 c747e19eb2667a3dca96a38a8dd562c5 HBM965.NVDC.564 2020-09-28             "Mul…
#> 14 e9ef358a160a32294aa8dbe3f3c09297 HBM237.XGGL.844 2020-09-28             "Mul…
#> 15 117c31327191f720e26b901e58a1f68b HBM284.FSBD.955 2020-09-28             "Mul…
#> 16 36a64b1b9a02b11e3f55ea84e6d3203a HBM427.DBBN.329 2020-09-28             "Mul…
#> 17 381f65e58d5e2c1d16a9cef2cc203aab HBM289.PCBM.487 2020-09-28             "Mul…
#> 18 6006923624c4745cbee806a331aaff54 HBM273.THFN.643 2020-09-28             "Mul…
#> # ℹ abbreviated name: ¹​last_modified_timestamp
collections_default_columns()
#> # A tibble: 4 × 1
#>   columns                
#>   <chr>                  
#> 1 uuid                   
#> 2 hubmap_id              
#> 3 title                  
#> 4 last_modified_timestamp

uuid <- "a320abcf56241be2e5e453401efeaaf3"
collection_detail(uuid)
#> # A tibble: 1 × 17
#>   contacts   contributors created_by_user_displayname created_by_user_email
#>   <list>     <list>       <chr>                       <chr>                
#> 1 <list [1]> <list [16]>  Bill Shirey                 shirey@pitt.edu      
#> # ℹ 13 more variables: created_timestamp <dbl>, datasets <list>,
#> #   description <chr>, doi_url <chr>, entity_type <chr>, group_name <chr>,
#> #   group_uuid <chr>, hubmap_id <chr>, index_version <chr>,
#> #   last_modified_timestamp <dbl>, registered_doi <chr>, title <chr>,
#> #   uuid <chr>
uuid <- "381f65e58d5e2c1d16a9cef2cc203aab"
collection_contacts(uuid)
#> # A tibble: 1 × 3
#>   name                 affiliation           orcid_id           
#>   <chr>                <chr>                 <chr>              
#> 1 Jeffrey M. Spraggins Vanderbilt University 0000-0001-9198-5498
uuid <- "381f65e58d5e2c1d16a9cef2cc203aab"
collection_data(uuid)
#> # A tibble: 4 × 7
#>   uuid      hubmap_id dataset_type_additio…¹ dataset_type last_modified_timest…²
#>   <chr>     <chr>     <chr>                  <chr>        <chr>                 
#> 1 a11fe9d1… HBM592.J… AF                     Auto-fluore… 2022-02-12            
#> 2 6e7ba37f… HBM525.D… MALDI-IMS-pos          MALDI        2022-02-12            
#> 3 484b359a… HBM264.X… PAS                    Histology    2024-08-08            
#> 4 60ed8e03… HBM253.W… MALDI-IMS-neg          MALDI        2022-02-12            
#> # ℹ abbreviated names: ¹​dataset_type_additional_information,
#> #   ²​last_modified_timestamp
#> # ℹ 2 more variables: status <chr>, organ <chr>
uuid <- "590b0485a196956284b8f3344276bc50"
collection_contributors(uuid)
#> # A tibble: 16 × 3
#>    name                   affiliation                    orcid_id           
#>    <chr>                  <chr>                          <chr>              
#>  1 Elizabeth K. Neumann   Vanderbilt University          0000-0002-6078-3321
#>  2 Nathan Heath Patterson Vanderbilt University          0000-0002-0064-1583
#>  3 David M.G. Anderson    Vanderbilt University          0000-0002-3866-0923
#>  4 Kavya Sharman          Vanderbilt University          0000-0002-3487-7199
#>  5 Lukasz Migas           Delft University of Technology 0000-0002-1884-6405
#>  6 Jamie L. Allen         Vanderbilt University          0000-0002-4739-2166
#>  7 Maya Brewer            Vanderbilt University          0000-0001-5914-0081
#>  8 Jennifer Harvey        Vanderbilt University          0000-0003-3067-1238
#>  9 Haichun Yang           Vanderbilt University          0000-0003-4265-7492
#> 10 Raymond C. Harris      Vanderbilt University          0000-0001-8025-0883
#> 11 Agnes B. Fogo          Vanderbilt University          0000-0003-3698-8527
#> 12 Mark deCaestecker      Vanderbilt University          0000-0001-7926-1673
#> 13 Danielle B. Gutierrez  Vanderbilt University          0000-0001-6355-2134
#> 14 Richard M. Caprioli    Vanderbilt University          0000-0001-5859-3310
#> 15 Raf Van de Plas        Delft University of Technology 0000-0002-2232-7130
#> 16 Jeffrey M. Spraggins   Vanderbilt University          0000-0001-9198-5498