HuBMAP Collections

collections() returns details about available collections, ordered by last modified dates

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

collection_contacts() takes a unique collection_id and returns contacts information of one specified collection as a tibble

collection_data() takes a unique collection_id and returns related datasets of one specified collection as a tibble

collection_contributors() takes a unique collection_id and returns contributors information of one specified collection as a tibble

Usage

collections(size = 10000L, from = 0L)

collections_default_columns(as = c("tibble", "character"))

collection_contacts(uuid)

collection_data(uuid)

collection_contributors(uuid)

Arguments

size

integer(1) number of maximum results to return; The default (10000) is meant to be large enough to return all results.

from

integer(1) number of number of results to skip, defaulting to 0.

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Collection UUID string. This is expected to be a 32-digit hex number.

Value

*_columns() returns a named list name

containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

collections(size = 4, from = 2)
#> # A tibble: 4 × 4
#>   uuid                             hubmap_id       last_modified_timestamp title
#>   <chr>                            <chr>           <chr>                   <chr>
#> 1 c87fa4ec3aa87af19a9ad29d2e2b4c5c HBM776.RGSW.867 2022-09-02              "\"A…
#> 2 34b068d4a926f77fd98b3d968b6c172f HBM426.WQXQ.244 2022-04-20              "Hum…
#> 3 4964d24bbc6668a72c4cbb5e0393a6bc HBM925.SGXL.596 2021-05-19              "HuB…
#> 4 6a6efd0c1a2681dc7d2faab8e4ab0bca HBM876.XNRH.336 2020-09-28              "Mul…
collections_default_columns()
#> # A tibble: 4 × 1
#>   columns                
#>   <chr>                  
#> 1 uuid                   
#> 2 hubmap_id              
#> 3 title                  
#> 4 last_modified_timestamp

uuid <- "381f65e58d5e2c1d16a9cef2cc203aab"
collection_contacts(uuid)
#> # A tibble: 1 × 3
#>   name                 affiliation           orcid_id           
#>   <chr>                <chr>                 <chr>              
#> 1 Jeffrey M. Spraggins Vanderbilt University 0000-0001-9198-5498
uuid <- "381f65e58d5e2c1d16a9cef2cc203aab"
collection_data(uuid)
#> # A tibble: 4 × 7
#>   uuid     hubmap_id data_types dataset_type last_modified_timest…¹ status organ
#>   <chr>    <chr>     <chr>      <chr>        <chr>                  <chr>  <chr>
#> 1 a11fe9d… HBM592.J… AF         Auto-fluore… 2022-02-12             Publi… Kidn…
#> 2 6e7ba37… HBM525.D… MALDI-IMS… MALDI        2022-02-12             Publi… Kidn…
#> 3 484b359… HBM264.X… PAS        Histology    2022-02-12             Publi… Kidn…
#> 4 60ed8e0… HBM253.W… MALDI-IMS… MALDI        2022-02-12             Publi… Kidn…
#> # ℹ abbreviated name: ¹​last_modified_timestamp
uuid <- "590b0485a196956284b8f3344276bc50"
collection_contributors(uuid)
#> # A tibble: 16 × 3
#>    name                   affiliation                    orcid_id           
#>    <chr>                  <chr>                          <chr>              
#>  1 Elizabeth K. Neumann   Vanderbilt University          0000-0002-6078-3321
#>  2 Nathan Heath Patterson Vanderbilt University          0000-0002-0064-1583
#>  3 David M.G. Anderson    Vanderbilt University          0000-0002-3866-0923
#>  4 Kavya Sharman          Vanderbilt University          0000-0002-3487-7199
#>  5 Lukasz Migas           Delft University of Technology 0000-0002-1884-6405
#>  6 Jamie L. Allen         Vanderbilt University          0000-0002-4739-2166
#>  7 Maya Brewer            Vanderbilt University          0000-0001-5914-0081
#>  8 Jennifer Harvey        Vanderbilt University          0000-0003-3067-1238
#>  9 Haichun Yang           Vanderbilt University          0000-0003-4265-7492
#> 10 Raymond C. Harris      Vanderbilt University          0000-0001-8025-0883
#> 11 Agnes B. Fogo          Vanderbilt University          0000-0003-3698-8527
#> 12 Mark deCaestecker      Vanderbilt University          0000-0001-7926-1673
#> 13 Danielle B. Gutierrez  Vanderbilt University          0000-0001-6355-2134
#> 14 Richard M. Caprioli    Vanderbilt University          0000-0001-5859-3310
#> 15 Raf Van de Plas        Delft University of Technology 0000-0002-2232-7130
#> 16 Jeffrey M. Spraggins   Vanderbilt University          0000-0001-9198-5498