HuBMAP Datasets

datasets returns the details available datasets, ordered by last modified dates

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

dataset_detail() takes a unique dataset_id and returns details about one specified dataset as a tibble

dataset_derived() takes a unique sample_id and returns the derived (support) dataset details. Support datasets normally belong to Image Pyramid, with image files available to download via Globus Collection. See details to download in files_globus_url().

Usage

datasets()

datasets_default_columns(as = c("tibble", "character"))

dataset_detail(uuid)

dataset_derived(uuid)

Arguments

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Sample UUID string. This is expected to be a 32-digit hex number.

Value

*_columns() returns a named list name

containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

datasets()
#> # A tibble: 2,889 × 13
#>    uuid              hubmap_id registered_by dataset_processing_c…¹ dataset_type
#>    <chr>             <chr>     <chr>         <chr>                  <chr>       
#>  1 3e4c568d9ce8df9d… HBM422.K… HuBMAP Proce… Raw                    MUSIC       
#>  2 db0d7736f53dcf9b… HBM762.L… HuBMAP Proce… HuBMAP Process         Histology […
#>  3 baf976734dd65220… HBM468.M… HuBMAP Proce… Raw                    Histology   
#>  4 bbbf5a5b29986dd5… HBM664.N… HuBMAP Proce… HuBMAP Process         Histology […
#>  5 57b475b8ad39fa2c… HBM579.L… HuBMAP Proce… Raw                    Histology   
#>  6 856aa4e05b68a784… HBM433.H… HuBMAP Proce… Raw                    Histology   
#>  7 a86f65547dbb2528… HBM989.N… HuBMAP Proce… Raw                    MUSIC       
#>  8 bc88a7cec9952a4c… HBM436.Z… HuBMAP Proce… Raw                    MUSIC       
#>  9 8a7d2e1f06f01754… HBM634.F… HuBMAP Proce… Raw                    MUSIC       
#> 10 80ec78509e49ec71… HBM644.W… HuBMAP Proce… Raw                    MUSIC       
#> # ℹ 2,879 more rows
#> # ℹ abbreviated name: ¹​dataset_processing_category
#> # ℹ 8 more variables: donor.hubmap_id <chr>, group_name <chr>,
#> #   last_modified_timestamp <chr>, analyte_class <chr>, sample_category <chr>,
#> #   status <chr>, data_types <chr>, organ <chr>
datasets_default_columns()
#> # A tibble: 13 × 1
#>    columns                    
#>    <chr>                      
#>  1 uuid                       
#>  2 hubmap_id                  
#>  3 group_name                 
#>  4 data_types                 
#>  5 dataset_type               
#>  6 organ                      
#>  7 analyte_class              
#>  8 dataset_processing_category
#>  9 sample_category            
#> 10 registered_by              
#> 11 status                     
#> 12 last_modified_timestamp    
#> 13 donor.hubmap_id            

uuid <- "7754aa5ebde628b5e92705e33e74a4ef"
dataset_detail(uuid)
#> # A tibble: 1 × 31
#>   ancestor_ids ancestors  contains_human_genetic_sequen…¹ created_by_user_disp…²
#>   <list>       <list>     <lgl>                           <chr>                 
#> 1 <chr [4]>    <list [4]> FALSE                           HuBMAP Process        
#> # ℹ abbreviated names: ¹​contains_human_genetic_sequences,
#> #   ²​created_by_user_displayname
#> # ℹ 27 more variables: created_by_user_email <chr>, created_timestamp <dbl>,
#> #   creation_action <chr>, data_access_level <chr>, data_types <list>,
#> #   dataset_info <chr>, dataset_type <chr>, descendant_ids <list>,
#> #   descendants <list>, display_subtype <chr>, donor <list>, entity_type <chr>,
#> #   files <list>, group_name <chr>, group_uuid <chr>, hubmap_id <chr>, …
# no derived dataset
uuid <- "3acdb3ed962b2087fbe325514b098101"
dataset_derived(uuid)
#> NULL

# with derived dataset
uuid <- "2c77b1cdf33dbed3dbfb74e4b578300e"
dataset_derived(uuid)
#> # A tibble: 1 × 6
#>   uuid           hubmap_id data_types dataset_type status last_modified_timest…¹
#>   <chr>          <chr>     <chr>      <chr>        <chr>  <chr>                 
#> 1 9e7b040f23124… HBM896.G… salmon_rn… RNAseq [Sal… Publi… 2020-08-22            
#> # ℹ abbreviated name: ¹​last_modified_timestamp