HuBMAP Datasets

datasets returns the details available datasets, ordered by last modified dates

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

dataset_detail() takes a unique dataset_id and returns details about one specified dataset as a tibble

dataset_derived() takes a unique sample_id and returns the derived (support) dataset details. Support datasets normally belong to Image Pyramid, with image files available to download via Globus Collection. See details to download in files_globus_url().

dataset_metadata() takes a unique dataset_id and returns the metadata of the dataset.

dataset_contributors() takes a unique dataset_id and returns the contributors of the dataset. For questions for this dataset, reach out to the individuals listed as contacts, either via the email address listed in the table or contact information provided on their ORCID profile page.

Usage

datasets()

datasets_default_columns(as = c("tibble", "character"))

dataset_detail(uuid)

dataset_derived(uuid)

dataset_metadata(uuid)

dataset_contributors(uuid)

Arguments

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Donor UUID string. This is expected to be a 32-digit hex number.

Value

*_columns() returns a named list name containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

datasets()
#> # A tibble: 3,301 × 14
#>    uuid    hubmap_id dataset_type data_types organ analyte_class sample_category
#>    <chr>   <chr>     <chr>        <chr>      <chr> <chr>         <chr>          
#>  1 5904bf… HBM856.B… ATACseq      ""         Uter… DNA           suspension     
#>  2 bf399e… HBM634.F… ATACseq      ""         Uter… DNA           suspension     
#>  3 223c79… HBM625.W… RNAseq       ""         Uter… RNA           suspension     
#>  4 2da752… HBM647.G… ATACseq      ""         Fall… DNA           suspension     
#>  5 1991b9… HBM853.X… ATACseq      ""         Uter… DNA           suspension     
#>  6 dda3b0… HBM834.K… RNAseq       ""         Uter… RNA           suspension     
#>  7 0955bb… HBM249.L… ATACseq      ""         Uter… DNA           suspension     
#>  8 d7ef8d… HBM484.P… ATACseq      ""         Ovar… DNA           block          
#>  9 aaed24… HBM956.C… ATACseq      ""         Ovar… DNA           block          
#> 10 33f5d0… HBM999.W… RNAseq       ""         Fall… RNA           block          
#> # ℹ 3,291 more rows
#> # ℹ 7 more variables: status <chr>, dataset_processing_category <chr>,
#> #   pipeline <chr>, registered_by <chr>, donor.hubmap_id <chr>,
#> #   group_name <chr>, last_modified_timestamp <chr>
datasets_default_columns()
#> # A tibble: 14 × 1
#>    columns                    
#>    <chr>                      
#>  1 uuid                       
#>  2 hubmap_id                  
#>  3 group_name                 
#>  4 data_types                 
#>  5 dataset_type               
#>  6 organ                      
#>  7 analyte_class              
#>  8 dataset_processing_category
#>  9 sample_category            
#> 10 registered_by              
#> 11 status                     
#> 12 pipeline                   
#> 13 last_modified_timestamp    
#> 14 donor.hubmap_id            

uuid <- "7754aa5ebde628b5e92705e33e74a4ef"
dataset_detail(uuid)
#> # A tibble: 1 × 31
#>   ancestor_ids ancestors  contains_human_genetic_sequen…¹ created_by_user_disp…²
#>   <list>       <list>     <lgl>                           <chr>                 
#> 1 <chr [4]>    <list [4]> FALSE                           HuBMAP Process        
#> # ℹ abbreviated names: ¹​contains_human_genetic_sequences,
#> #   ²​created_by_user_displayname
#> # ℹ 27 more variables: created_by_user_email <chr>, created_timestamp <dbl>,
#> #   creation_action <chr>, data_access_level <chr>, data_types <list>,
#> #   dataset_info <chr>, dataset_type <chr>, descendant_ids <list>,
#> #   descendants <list>, display_subtype <chr>, donor <list>, entity_type <chr>,
#> #   files <list>, group_name <chr>, group_uuid <chr>, hubmap_id <chr>, …
# no derived dataset
uuid <- "3acdb3ed962b2087fbe325514b098101"
dataset_derived(uuid)
#> NULL

# with derived dataset
uuid <- "2c77b1cdf33dbed3dbfb74e4b578300e"
dataset_derived(uuid)
#> # A tibble: 1 × 6
#>   uuid           hubmap_id data_types dataset_type status last_modified_timest…¹
#>   <chr>          <chr>     <chr>      <chr>        <chr>  <chr>                 
#> 1 9e7b040f23124… HBM896.G… salmon_rn… RNAseq [Sal… Publi… 2020-08-22            
#> # ℹ abbreviated name: ¹​last_modified_timestamp
uuid <- "564167adbbb2fdd64c24e7ea409c23f1"
dataset_metadata(uuid)
#> # A tibble: 43 × 2
#>    Key                             Value                  
#>    <chr>                           <chr>                  
#>  1 acquisition_instrument_model    BZ-X710                
#>  2 acquisition_instrument_vendor   Keyence                
#>  3 analyte_class                   Polysaccharide         
#>  4 contributors_path               extras/contributors.tsv
#>  5 data_path                       .                      
#>  6 dataset_type                    Histology              
#>  7 is_batch_staining_done          No                     
#>  8 is_image_preprocessing_required No                     
#>  9 is_staining_automated           No                     
#> 10 is_targeted                     No                     
#> # ℹ 33 more rows

uuid <- "564167adbbb2fdd64c24e7ea409c23f1"
dataset_contributors(uuid)
#> # A tibble: 3 × 5
#>   display_name    affiliation email                 orcid is_principal_investi…¹
#>   <chr>           <chr>       <chr>                 <chr> <chr>                 
#> 1 John Hickey     Stanford    john.hickey@duke.edu  0000… No                    
#> 2 Chiara Caraccio Stanford    caraccio@stanford.edu 0000… No                    
#> 3 Garry Nolan     Stanford    gnolan@stanford.edu   0000… Yes                   
#> # ℹ abbreviated name: ¹​is_principal_investigator