HuBMAP Donors

donors() returns details about available samples, ordered by last modified dates

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

donor_detail() takes a unique donor_id and returns details about one specified sample as a tibble

donor_derived() takes a unique donor_id and returns the derived dataset or/and sample details.

donor_metadata() takes a unique donor_id and returns the metadata of the donor.

Usage

donors()

donors_default_columns(as = c("tibble", "character"))

donor_detail(uuid)

donor_derived(uuid, entity_type = c("Dataset", "Sample"))

donor_metadata(uuid)

Arguments

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Donor UUID string. This is expected to be a 32-digit hex number.

entity_type

character(1) selected derived entity type. One of "Sample" or "Dataset" (default).

Value

*_columns() returns a named list name

containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

donors()
#> # A tibble: 228 × 8
#>    uuid                 hubmap_id group_name Sex   Age   `Body Mass Index` Race 
#>    <chr>                <chr>     <chr>      <chr> <chr> <chr>             <chr>
#>  1 c09f875545a64694d70… HBM534.P… TTD - Uni… Fema… 90    NA                White
#>  2 0abacde2443881351ff… HBM358.K… TTD - Uni… Male  90    NA                White
#>  3 d7c90f1fe1ae702df84… HBM639.M… TTD - Uni… Fema… 59    NA                White
#>  4 b1acd1731563347de3a… HBM674.F… TTD - Uni… Fema… 60    NA                White
#>  5 24e6946400b65468a96… HBM379.B… TTD - Uni… Male  79    NA                White
#>  6 c9c39035edc9de812f6… HBM543.J… TTD - Uni… Male  71    NA                White
#>  7 f079501ef0915131c8b… HBM448.V… TTD - Uni… Fema… 82    NA                White
#>  8 a0bea4ae0ce4e03efd0… HBM666.X… TTD - Uni… Fema… 82    NA                White
#>  9 03b6236566cfd447214… HBM263.S… TTD - Uni… Fema… 75    NA                White
#> 10 5cc62ac9233e8c5d6c4… HBM993.R… TTD - Uni… Fema… 75    NA                White
#> # ℹ 218 more rows
#> # ℹ 1 more variable: last_modified_timestamp <chr>
donors_default_columns()
#> # A tibble: 8 × 1
#>   columns                
#>   <chr>                  
#> 1 hubmap_id              
#> 2 uuid                   
#> 3 group_name             
#> 4 Sex                    
#> 5 Age                    
#> 6 Body Mass Index        
#> 7 Race                   
#> 8 last_modified_timestamp

uuid <- "1dcde05aea3509b2cf89a41ceb3d700f"
donor_detail(uuid)
#> # A tibble: 1 × 24
#>   ancestor_ids ancestors  created_by_user_displayname created_by_user_email     
#>   <list>       <list>     <chr>                       <chr>                     
#> 1 <list [0]>   <list [0]> Jamie Allen                 jamie.l.allen@vanderbilt.…
#> # ℹ 20 more variables: created_timestamp <dbl>, data_access_level <chr>,
#> #   descendant_ids <list>, descendants <list>, description <chr>,
#> #   display_subtype <chr>, entity_type <chr>, group_name <chr>,
#> #   group_uuid <chr>, hubmap_id <chr>, immediate_ancestors <list>,
#> #   immediate_descendants <list>, index_version <chr>, lab_donor_id <chr>,
#> #   label <chr>, last_modified_timestamp <dbl>, metadata <list>,
#> #   protocol_url <chr>, submission_id <chr>, uuid <chr>
uuid <- "1dcde05aea3509b2cf89a41ceb3d700f"
donor_derived(uuid, "Sample")
#> # A tibble: 5 × 6
#>   uuid                    hubmap_id sample_category last_modified_timest…¹ organ
#>   <chr>                   <chr>     <chr>           <chr>                  <chr>
#> 1 146c221bca3759ffa64cd0… HBM847.D… organ           2021-08-04             Kidn…
#> 2 27485b21a2285dbcae38f2… HBM693.K… block           2022-07-19             Kidn…
#> 3 1441346ca08aadbb631539… HBM488.F… section         2021-08-25             Kidn…
#> 4 e97fd3e2b6dae2238cfe05… HBM695.M… section         2021-08-25             Kidn…
#> 5 92cd878bf5be78131c6a3f… HBM686.B… section         2021-08-25             Kidn…
#> # ℹ abbreviated name: ¹​last_modified_timestamp
#> # ℹ 1 more variable: derived_dataset_count <int>

uuid <- "d37df2cad4e80dc368763caefccf7140"
donor_metadata(uuid)
#> # A tibble: 8 × 2
#>   Key             Value           
#>   <chr>           <chr>           
#> 1 Age             "49 year"       
#> 2 Body Mass Index "31.2 kg/m2"    
#> 3 Cause of Death  "Gunshot wound "
#> 4 Death Event     "Suicide "      
#> 5 Height          "72 inches"     
#> 6 Race            "White "        
#> 7 Sex             "Male "         
#> 8 Weight          "230 pounds"