HuBMAP Samples

publications() returns details about available samples, ordered by last modified dates. publication_status == TRUE: peer reviewed publication; publication_status == FALSE: pre-print publication.

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

publication_data() takes a unique publication_id and returns details about one specified publication.

publication_authors() takes a unique publication_id and returns authors information of one specified collection as a tibble

Usage

publications()

publications_default_columns(as = c("tibble", "character"))

publication_data(uuid, entity_type = c("Dataset", "Sample", "Donor"))

publication_authors(uuid)

Arguments

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Publication UUID string. This is expected to be a 32-digit hex number.

entity_type

character(1) selected derived entity type. One of "Sample", "Donor" or "Dataset" (default).

Value

*_columns() returns a named list name containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

publications()
#> # A tibble: 9 × 7
#>   uuid      hubmap_id last_modified_timest…¹ publication_date publication_status
#>   <chr>     <chr>     <chr>                  <chr>            <chr>             
#> 1 77ab3588… HBM222.S… 2023-07-18             2023-07-20       TRUE              
#> 2 aea510e3… HBM332.M… 2023-07-18             2022-10-24       TRUE              
#> 3 2ced91fd… HBM885.X… 2023-07-18             2021-10-18       FALSE             
#> 4 e8338966… HBM695.W… 2023-07-18             2023-07-19       TRUE              
#> 5 dd0e7aea… HBM599.D… 2023-07-18             2023-07-19       TRUE              
#> 6 3c727366… HBM629.D… 2023-07-18             2023-07-19       TRUE              
#> 7 0de8528c… HBM473.R… 2023-07-18             2023-07-19       TRUE              
#> 8 eb975811… HBM464.W… 2023-07-18             2023-07-19       TRUE              
#> 9 72cbeb8f… HBM222.Z… 2023-07-18             2023-07-19       TRUE              
#> # ℹ abbreviated name: ¹​last_modified_timestamp
#> # ℹ 2 more variables: publication_venue <chr>, title <chr>
publications_default_columns()
#> # A tibble: 7 × 1
#>   columns                
#>   <chr>                  
#> 1 uuid                   
#> 2 hubmap_id              
#> 3 title                  
#> 4 publication_venue      
#> 5 last_modified_timestamp
#> 6 publication_date       
#> 7 publication_status     

uuid <- "3c7273660cdf9ab91a7901533b2cd9a5"
publication_data(uuid)
#> # A tibble: 1 × 7
#>   uuid     hubmap_id data_types dataset_type status last_modified_timest…¹ organ
#>   <chr>    <chr>     <chr>      <chr>        <chr>  <chr>                  <chr>
#> 1 00d9d4b… HBM342.M… SNARE-RNA… RNAseq       Publi… 2022-05-11             Kidn…
#> # ℹ abbreviated name: ¹​last_modified_timestamp

uuid <- "3c7273660cdf9ab91a7901533b2cd9a5"
publication_authors(uuid)
#> # A tibble: 50 × 3
#>    name                  affiliation                                    orcid_id
#>    <chr>                 <chr>                                          <chr>   
#>  1 Blue B. Lake          Department of Bioengineering, University of C… 0000-00…
#>  2 Rajasree Menon        Department of Computational Medicine and Bioi… 0000-00…
#>  3 Seth Winfree          Department of Pathology and Microbiology, Uni… 0000-00…
#>  4 Qiwen Hu              Department of Biomedical Informatics, Harvard… 0000-00…
#>  5 Ricardo Melo Ferreira Department of Medicine, Indiana University Sc… 0000-00…
#>  6 Kian Kalhor           Department of Bioengineering, University of C… 0000-00…
#>  7 Daria Barwinska       Department of Medicine, Indiana University Sc… 0000-00…
#>  8 Edgar A. Otto         Department of Internal Medicine, Division of … 0000-00…
#>  9 Michael Ferkowicz     Department of Medicine, Indiana University Sc… 0000-00…
#> 10 Dinh Diep             Department of Bioengineering, University of C… 0000-00…
#> # ℹ 40 more rows