HuBMAP Samples

samples() returns details about available samples, ordered by last modified dates

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

sample_detail() takes a unique sample_id and returns details about one specified sample as a tibble

sample_derived() takes a unique sample_id and returns the derived dataset or/and sample details.

Usage

samples(size = 10000L, from = 0L)

samples_default_columns(as = c("tibble", "character"))

sample_detail(uuid)

sample_derived(uuid, entity_type = c("Dataset", "Sample"))

Arguments

size

integer(1) number of maximum results to return; The default (10000) is meant to be large enough to return all results.

from

integer(1) number of number of results to skip, defaulting to 0.

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Sample UUID string. This is expected to be a 32-digit hex number.

entity_type

character(1) selected derived entity type. One of "Sample" or "Dataset".

Value

*_columns() returns a named list name containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

samples(size = 20, from = 5)
#> # A tibble: 20 × 7
#>    uuid              hubmap_id donor.hubmap_id group_name last_modified_timest…¹
#>    <chr>             <chr>     <chr>           <chr>      <chr>                 
#>  1 1cfe67afe97094ac… HBM355.M… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  2 038e4fcefc00f7ec… HBM595.S… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  3 23820292ab47f319… HBM236.X… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  4 e3b74d607c3b20c9… HBM876.S… HBM477.DBJN.952 Universit… 2024-05-29            
#>  5 e4ffe5a1de1e63b3… HBM867.F… HBM568.HLJW.252 Universit… 2024-05-29            
#>  6 744f20be06ec72b7… HBM796.X… HBM375.PJSC.682 Universit… 2024-05-29            
#>  7 681f7f1d0a87427c… HBM788.K… HBM984.MHHN.738 Universit… 2024-05-29            
#>  8 a6acb120173be820… HBM786.T… HBM662.VVHK.499 Universit… 2024-05-29            
#>  9 0ec930e0f4a8fcf1… HBM752.D… HBM994.SGLP.263 Universit… 2024-05-29            
#> 10 acb712bf26d8fe7d… HBM748.G… HBM735.QDFT.264 Universit… 2024-05-29            
#> 11 c18dbe465fe2dfdd… HBM735.Q… HBM894.PKMC.242 Universit… 2024-05-29            
#> 12 5abbecf8d13b9df7… HBM562.L… HBM256.BRQS.425 Universit… 2024-05-29            
#> 13 8d7ce3aab7013e41… HBM556.C… HBM867.BLQJ.787 Universit… 2024-05-29            
#> 14 c4200fa2bd88eca8… HBM486.Q… HBM984.MHHN.738 Universit… 2024-05-29            
#> 15 c15e6342d6905d6d… HBM548.T… HBM579.CTRK.363 Universit… 2024-05-29            
#> 16 d9e64e77d26d9fb8… HBM455.P… HBM377.WFTZ.879 Universit… 2024-05-29            
#> 17 2e398e91f1336b67… HBM387.K… HBM386.FNTC.294 Universit… 2024-05-29            
#> 18 8cfd588e5bc4a31e… HBM865.P… HBM892.LCXT.493 General E… 2024-05-29            
#> 19 f187ff9a4631540a… HBM297.V… HBM372.JZTV.423 General E… 2024-05-29            
#> 20 c6f7a9e8189e2111… HBM393.S… HBM593.NGPZ.869 TTD - Uni… 2024-05-14            
#> # ℹ abbreviated name: ¹​last_modified_timestamp
#> # ℹ 2 more variables: sample_category <chr>, organ <chr>
samples_default_columns()
#> # A tibble: 7 × 1
#>   columns                
#>   <chr>                  
#> 1 uuid                   
#> 2 hubmap_id              
#> 3 group_name             
#> 4 sample_category        
#> 5 organ                  
#> 6 last_modified_timestamp
#> 7 donor.hubmap_id        

uuid <- "d3525d35f6d5ee3dc3186613b0ab1762"
sample_detail(uuid)
#> # A tibble: 1 × 25
#>   ancestor_ids ancestors  created_by_user_displayname created_by_user_email
#>   <list>       <list>     <chr>                       <chr>                
#> 1 <chr [3]>    <list [3]> Marda Jorgensen             marda@ufl.edu        
#> # ℹ 21 more variables: created_timestamp <dbl>, data_access_level <chr>,
#> #   descendant_ids <list>, descendants <list>, display_subtype <chr>,
#> #   donor <list>, entity_type <chr>, group_name <chr>, group_uuid <chr>,
#> #   hubmap_id <chr>, immediate_ancestors <list>, immediate_descendants <list>,
#> #   index_version <chr>, lab_tissue_sample_id <chr>,
#> #   last_modified_timestamp <dbl>, origin_samples <list>, protocol_url <chr>,
#> #   rui_location <chr>, sample_category <chr>, submission_id <chr>, …
uuid <- "3e7dc14313262af577f686dcb09f5119"
sample_derived(uuid, "Sample")
#> # A tibble: 1 × 6
#>   uuid         hubmap_id group_name sample_category last_modified_timest…¹ organ
#>   <chr>        <chr>     <chr>      <chr>           <chr>                  <chr>
#> 1 63ef1040ae3… HBM976.W… TTD - Uni… section         2024-04-09             Brain
#> # ℹ abbreviated name: ¹​last_modified_timestamp