HuBMAP Samples

samples() returns details about available samples, ordered by last modified dates. There are multiple tissue sample types displayed in sample_category column which are block, organ, suspension, and section.

*_columns() returns a tibble or named character vector describing the content of the tibble returned by samples(), datasets(), donors(), collections(), or publications().

sample_detail() takes a unique sample_id and returns details about one specified sample as a tibble

sample_derived() takes a unique sample_id and returns the derived dataset or/and sample details.

sample_metadata() takes a unique donor_id and returns the metadata of the sample.

Usage

samples()

samples_default_columns(as = c("tibble", "character"))

sample_detail(uuid)

sample_derived(uuid, entity_type = c("Dataset", "Sample"))

sample_metadata(uuid)

Arguments

as

character(1) return format. One of "tibble" (default), or "character".

uuid

character(1) corresponding to the HuBMAP Donor UUID string. This is expected to be a 32-digit hex number.

entity_type

character(1) selected derived entity type. One of "Sample" or "Dataset" (default).

Value

*_columns() returns a named list name containing the column name used in the tibble returned by samples(), datasets(), donors(), collections(), or publications(). When as = "tibble",the return value is a tibble with paths as elements and abbreviations as names.

Details

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium webpage, https://software.docs.hubmapconsortium.org/apis

Examples

samples()
#> # A tibble: 2,191 × 7
#>    uuid              hubmap_id donor_hubmap_id group_name last_modified_timest…¹
#>    <chr>             <chr>     <chr>           <chr>      <chr>                 
#>  1 5abbecf8d13b9df7… HBM562.L… HBM256.BRQS.425 Universit… 2024-10-17            
#>  2 c18dbe465fe2dfdd… HBM735.Q… HBM894.PKMC.242 Universit… 2024-10-17            
#>  3 c15e6342d6905d6d… HBM548.T… HBM579.CTRK.363 Universit… 2024-10-17            
#>  4 f7b37eff8a36eb38… HBM997.B… HBM765.TRHD.452 Stanford … 2024-09-23            
#>  5 44af15225914d2b3… HBM866.F… HBM963.LXWM.598 TC - Univ… 2024-07-30            
#>  6 569fbb861c163632… HBM955.D… HBM883.LHNK.872 TC - Univ… 2024-07-30            
#>  7 6874561561f77469… HBM435.T… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  8 6f700af74a252c99… HBM363.L… HBM279.WPZP.978 Stanford … 2024-07-19            
#>  9 f0905232556b1608… HBM396.V… HBM279.WPZP.978 Stanford … 2024-07-19            
#> 10 038e4fcefc00f7ec… HBM595.S… HBM279.WPZP.978 Stanford … 2024-07-19            
#> # ℹ 2,181 more rows
#> # ℹ abbreviated name: ¹​last_modified_timestamp
#> # ℹ 2 more variables: sample_category <chr>, organ <chr>
samples_default_columns()
#> # A tibble: 7 × 1
#>   columns                
#>   <chr>                  
#> 1 uuid                   
#> 2 hubmap_id              
#> 3 group_name             
#> 4 sample_category        
#> 5 organ                  
#> 6 last_modified_timestamp
#> 7 donor_hubmap_id        

uuid <- "d3525d35f6d5ee3dc3186613b0ab1762"
sample_detail(uuid)
#> # A tibble: 1 × 23
#>   ancestor_ids ancestors  created_by_user_displayname created_by_user_email
#>   <list>       <list>     <chr>                       <chr>                
#> 1 <chr [3]>    <list [3]> Marda Jorgensen             marda@ufl.edu        
#> # ℹ 19 more variables: created_timestamp <dbl>, data_access_level <chr>,
#> #   descendant_ids <list>, descendants <list>, display_subtype <chr>,
#> #   donor <list>, entity_type <chr>, group_name <chr>, group_uuid <chr>,
#> #   hubmap_id <chr>, immediate_ancestor_ids <list>,
#> #   immediate_descendant_ids <list>, index_version <chr>,
#> #   last_modified_timestamp <dbl>, origin_samples <list>, protocol_url <chr>,
#> #   rui_location <chr>, sample_category <chr>, uuid <chr>
uuid <- "3e7dc14313262af577f686dcb09f5119"
sample_derived(uuid, "Sample")
#> # A tibble: 1 × 2
#>   uuid                             organ
#>   <chr>                            <chr>
#> 1 63ef1040ae30473778dc4bd2f2749fe3 Brain
uuid <- "8d7ce3aab7013e416263d23b0c048900"
sample_metadata(uuid)
#> # A tibble: 9 × 2
#>   Key                   Value                                                   
#>   <chr>                 <chr>                                                   
#> 1 donor.Age             "54 years"                                              
#> 2 donor.Body Mass Index "53.06 kg/m2"                                           
#> 3 donor.Cause of Death  "Anoxia "                                               
#> 4 donor.Height          "157.48 cm"                                             
#> 5 donor.Medical History "Type 2 Diabetes Mellitus ; Hypertension ; Asthma ; Chr…
#> 6 donor.Race            "Black or African American "                            
#> 7 donor.Sex             "Female "                                               
#> 8 donor.Social History  "Smoker ; Illicit drug use "                            
#> 9 donor.Weight          "131.6 kg"